evaluators/aws/pyproject.toml
[tool.poetry]
name = "langevals-aws"
version = "0.1.0"
description = "LangEvals integration for AWS APIs evaluators"
authors = ["Rogerio Chaves <rogerio@langwatch.ai>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.11"
langevals-core = { path = "../../langevals_core", develop = true }
boto3 = "^1.34.67"

[tool.poetry.group.test.dependencies]
pytest = "^7.4.2"
python-dotenv = "1.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
